refactor(car): Car를 model 폴더로 이동
